requirements.txt
Flask==2.0.2
gunicorn==20.1.0
python-dotenv==0.15.0
google-cloud-bigquery==2.34.0
